requirements.txt
openai>=1.0.0
pypdf>=3.0.0
python-dotenv>=1.0.0
langchain>=0.0.267
sentence-transformers>=2.2.2
chromadb>=0.4.13
umap-learn>=0.5.3
matplotlib>=3.7.2
numpy>=1.24.0